+++ src/UniversalConvert.App/AppVersion.cs
@@ -1,4 +1,5 @@
 using System.Reflection;
+using UniversalConvert.Core;
 
 namespace UniversalConvert.App
 {