+++ src/UniversalConvert.App/UpdateChecker.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
+using UniversalConvert.Core;
 using UniversalConvert.Core.Diagnostics;
 
 namespace UniversalConvert.App